+++ src/answer/mod.rs
@@ -1,3 +1,4 @@
 mod precious;
 mod records;
+mod colorize;
 pub use precious::{get_answers, get_answers_with_client, SPLITTER};